Add Peter Lowe’s list to imported lists
The list at https://pgl.yoyo.org/adservers/ maintained by Peter Lowe has
been around for years and seems to be well maintained and respected. The
author also explicitly says to “Feel free to combine this list with
yours or lists from other sites and put it up on the web, though!”
which, well, is exactly what we’re doing here. :)